Add new undo redo button and subdivide script for better readability
