+++ src/fastga_he/gui/power_train_builder/power_train_builder_handler.py
@@ -10,6 +10,18 @@ the appropriate action: placing a component on the canvas, selecting a node
 for editing, deleting a node or edge, managing port connections, and saving
 or loading the canvas state.
 
+The class is assembled from four single-responsibility mixins:
+
+* :class:`~.power_train_builder_handler_overlay.OverlayMixin` –
+  startup overlay, new-design reset, delete mode, palette selection.
+* :class:`~.power_train_builder_handler_io.IOMixin` –
+  save / load (YAML + JSON), tkinter file dialogs, end-session triggers,
+  undo / redo snapshot management.
+* :class:`~.power_train_builder_handler_ports.PortEdgeMixin` –
+  port hit-testing, edge add / remove / rebuild, nearest-node helpers.
+* :class:`~.power_train_builder_handler_config_panel.ConfigPanelMixin` –
+  node-table, options table, connections table, symmetry select.
+
 Typical usage::
 
     palette_layout, table_panel, state = ComponentPaletteConfigurationTableBuilder.build()
@@ -20,38 +32,25 @@ Typical usage::
 import json
 import logging
 from pathlib import Path
-from tkinter import filedialog
-import tkinter as tk
-import bokeh.models as bkmodel
-from bokeh.layouts import row
-from tornado.ioloop import IOLoop
 
-from fastga_he.gui.power_train_network_viewer import (
-    ICONS_CONFIG,
-    _string_cleanup,
-    _url_to_base64,
-    DEFAULT_COLOR,
-)
-from .power_train_network_writer import PowerTrainYAML
-from .power_train_builder_state import (
-    BuilderState,
-    BUTTON_DEFAULT_COLOR_TYPE,
-    BUTTON_SELECTED_COLOR_TYPE,
-    NODE_RADIUS,
-    PORT_RADIUS,
-    _EMPTY,
-)
-from .power_train_builder_state import compute_ports
+from fastga_he.gui.power_train_network_viewer import ICONS_CONFIG, _url_to_base64
+
+from .power_train_builder_state import BuilderState, _EMPTY
+from .power_train_builder_history import UndoStack
+from .power_train_builder_handler_overlay import OverlayMixin
+from .power_train_builder_handler_io import IOMixin
+from .power_train_builder_handler_ports import PortEdgeMixin
+from .power_train_builder_handler_config_panel import ConfigPanelMixin
 
 _LOGGER = logging.getLogger(__name__)
 
 
-class PlacementHandler:
+class PlacementHandler(OverlayMixin, IOMixin, PortEdgeMixin, ConfigPanelMixin):
     """
     Wires palette button events to canvas placement and interaction events.
 
     Instantiate **after** building the palette; the constructor automatically
-    wires all button ``on_click`` callbacks::
+    wires all button ``on_click`` callbacks and initialises the undo stack::
 
         palette_layout, table_panel, state = ComponentPaletteConfigurationTableBuilder.build()
         handler = PlacementHandler(state, main_plot)
@@ -67,7 +66,7 @@ class PlacementHandler:
 
     def __init__(self, state: BuilderState, main_plot, icon_size: int = 50):
         """
-        Initialise the handler and wire all button callbacks.
+        Initialise the handler, set up the undo stack, and wire all button callbacks.
 
         :param state: Shared :class:`BuilderState` instance.
         :param main_plot: The Bokeh ``figure`` that acts as the placement canvas.
@@ -76,7 +75,14 @@ class PlacementHandler:
         self.state = state
         self.main_plot = main_plot
         self.icon_size = icon_size
+
+        # Initialise the undo stack on the shared state so all mixins can
+        # reach it via self.state.undo_stack.
+        self.state.undo_stack = UndoStack()
+
         self._wire_buttons()
+        # Set the initial enabled/disabled appearance of the undo/redo buttons.
+        self._refresh_undo_redo_buttons()
 
     # -----------------------------------------------------------------------
     # Internal wiring
@@ -95,6 +101,8 @@ class PlacementHandler:
         * The browse_load_trigger → :meth:`_on_browse_load` (tkinter open-file dialog).
         * The browse_save_trigger → :meth:`_on_browse_save` (tkinter save-file dialogs).
         * The browse_watcher_trigger → :meth:`_on_browse_watcher` (tkinter CSV path dialog).
+        * The Undo button → :meth:`_on_undo`.
+        * The Redo button → :meth:`_on_redo`.
         """
         # End Session is handled entirely via JS + end_session_trigger to avoid the
         # Python on_click firing unconditionally (before JS can gate on save state).
@@ -126,6 +134,13 @@ class PlacementHandler:
                 "value", self._on_end_session_save_trigger
             )
 
+        # Undo / redo buttons (created by the launcher and stored in state)
+        if self.state.undo_button is not None:
+            self.state.undo_button.on_click(self._on_undo)
+
+        if self.state.redo_button is not None:
+            self.state.redo_button.on_click(self._on_redo)
+
     def _make_select_callback(self, button_index: int):
         """
         Return a zero-argument closure that selects the component at *button_index*.
@@ -134,1738 +149,6 @@ class PlacementHandler:
         :return: Callback function suitable for wiring to a palette button.
         """
         return lambda: self.on_palette_select(button_index)
-
-    # -----------------------------------------------------------------------
-    # Startup overlay
-    # -----------------------------------------------------------------------
-
-    def _dismiss_startup_overlay(self):
-        """
-        Hide the startup overlay panel (and both its buttons).
-
-        Called as soon as the user makes a choice so the canvas is unobstructed.
-        """
-        if self.state.startup_overlay is not None:
-            self.state.startup_overlay.visible = False
-        if self.state.new_design_button is not None:
-            self.state.new_design_button.visible = False
-        if self.state.load_design_button is not None:
-            self.state.load_design_button.visible = False
-
-    def _on_new_design(self):
-        """
-        Start with a blank canvas.
-
-        Hides the startup overlay and resets every data source to empty so the
-        user begins with a completely fresh powertrain design.
-        """
-        self._dismiss_startup_overlay()
-
-        self.state.placed_nodes_source.data = {
-            "x": [],
-            "y": [],
-            "url": [],
-            "w": [],
-            "h": [],
-            "name": [],
-            "node_type": [],
-            "icon_type": [],
-            "position": [],
-            "options": [],
-            "n_sources": [],
-            "n_targets": [],
-            "symmetry_name": [],
-            "symmetry_node_index": [],
-        }
-        self.state.hover_source.data = dict(x=[], y=[], name=[], node_type=[])
-        self.state.edge_source.data = dict(
-            xs=[],
-            ys=[],
-            color=[],
-            starting_node_index=[],
-            starting_port_label=[],
-            starting_port_kind=[],
-            ending_node_index=[],
-            ending_port_label=[],
-            ending_port_kind=[],
-        )
-        self.state.source_port_source.data = dict(
-            x=[],
-            y=[],
-            color=[],
-            label=[],
-            node_index=[],
-            node_name=[],
-            node_type=[],
-            fill_alpha=[],
-            line_alpha=[],
-            connected=[],
-        )
-        self.state.target_port_source.data = dict(
-            x=[],
-            y=[],
-            color=[],
-            label=[],
-            node_index=[],
-            node_name=[],
-            node_type=[],
-            fill_alpha=[],
-            line_alpha=[],
-            connected=[],
-        )
-        self._clear_temp_edges()
-        self.state.placed_counter.clear()
-        self.state.selected_node_index = None
-        self.state.selected_component = None
-        self.state.delete_mode = False
-        self._clear_node_table()
-
-        if self.state.save_button is not None:
-            self.state.save_button.button_type = "success"
-
-        _LOGGER.info("New Design started – canvas cleared.")
-
-    # -----------------------------------------------------------------------
-    # Delete mode
-    # -----------------------------------------------------------------------
-
-    def _toggle_delete_mode(self):
-        """
-        Toggle delete mode on / off, updating button styling and status text.
-
-        When entering delete mode any active component selection and pending
-        port connection are cleared, the config panel is hidden, the Delete
-        button turns red, and the status label prompts the user to click a
-        node or edge to remove it.  Leaving delete mode restores the defaults.
-        """
-        self._cancel_pending_connection()
-        self.state.delete_mode = not self.state.delete_mode
-        if self.state.delete_mode:
-            self.state.selected_component = None
-            self.state.selected_node_index = None
-            self._clear_temp_edges()
-            self._clear_node_table()
-            for btn in self.state.buttons:
-                btn.button_type = BUTTON_DEFAULT_COLOR_TYPE
-            self.state.status_div.text = (
-                "<b style='color:#FF4444;font-size:14pt'>Delete mode: "
-                "click an icon / a connection to remove it</b>"
-            )
-            self.state.delete_button.button_type = "danger"
-        else:
-            self.state.status_div.text = (
-                "<i style='color:#aaa;font-size:14pt'>Select a component</i>"
-            )
-            self.state.delete_button.button_type = BUTTON_DEFAULT_COLOR_TYPE
-
-    # -----------------------------------------------------------------------
-    # Palette selection
-    # -----------------------------------------------------------------------
-
-    def on_palette_select(self, component_index: int):
-        """
-        Select the component at position *component_index* in :data:`ICONS_CONFIG`.
-
-        A second click on the same button deselects it. Updates button styling
-        and the status label accordingly. Can be called programmatically in
-        tests without a running server.
-
-        :param component_index: Zero-based index into ``list(ICONS_CONFIG.keys())``.
-        """
-        icon_keys = list(ICONS_CONFIG.keys())
-        if component_index < 0 or component_index >= len(icon_keys):
-            return
-
-        if self.state.selected_component == icon_keys[component_index]:
-            self.state.selected_component = None
-            self.state.buttons[component_index].button_type = BUTTON_DEFAULT_COLOR_TYPE
-            self.state.status_div.text = (
-                "<i style='color:#aaa;font-size:14pt'>Select a component</i>"
-            )
-            return
-
-        self._cancel_pending_connection()
-        self.state.selected_component = icon_keys[component_index]
-
-        if self.state.delete_mode:
-            self.state.delete_mode = False
-            if self.state.delete_button is not None:
-                self.state.delete_button.button_type = BUTTON_DEFAULT_COLOR_TYPE
-
-        for button_position, palette_button in enumerate(self.state.buttons):
-            palette_button.button_type = (
-                BUTTON_SELECTED_COLOR_TYPE
-                if button_position == component_index
-                else BUTTON_DEFAULT_COLOR_TYPE
-            )
-
-        label = _string_cleanup(self.state.selected_component)
-        self.state.status_div.text = f"<b style='color:#FFD700;font-size:14pt'>Placing: {label}</b>"
-
-    # -----------------------------------------------------------------------
-    # Save / load
-    # -----------------------------------------------------------------------
-
-    def _mark_unsaved(self):
-        """
-        Turn the Save button yellow to signal unsaved changes.
-
-        Called by every method that mutates the canvas.  The button returns
-        to green only after a successful save.
-        """
-        if self.state.save_button is not None:
-            self.state.save_button.button_type = "warning"
-
-    def _save_canvas_state(self, yaml_path: str = "", json_path: str = ""):
-        """
-        Serialise the current canvas to the file paths chosen by the user.
-
-        Used only by the ``prompt()`` fallback path where the browser cannot
-        write files directly and the user supplies a filesystem path manually.
-        Both parameters are optional – an empty string means that file is
-        skipped.
-
-        :param yaml_path: File name (or full path) for the YAML config.
-        :param json_path: File name (or full path) for the JSON backup.
-        """
-        nodes_data = {
-            key: list(values) for key, values in self.state.placed_nodes_source.data.items()
-        }
-        edges_data = {key: list(values) for key, values in self.state.edge_source.data.items()}
-        source_port_data = {
-            key: list(values) for key, values in self.state.source_port_source.data.items()
-        }
-        target_port_data = {
-            key: list(values) for key, values in self.state.target_port_source.data.items()
-        }
-
-        if yaml_path:
-            yaml_file = Path(yaml_path)
-            yaml_file.parent.mkdir(parents=True, exist_ok=True)
-            try:
-                pt_yaml = PowerTrainYAML(self.state)
-                pt_yaml.set_title(yaml_file.stem)
-                for node_index in range(len(nodes_data.get("name", []))):
-                    pt_yaml.add_component(node_index)
-                pt_yaml.add_connection()
-                # Apply watcher file path when the user supplied one in the overlay.
-                watcher_path = (
-                    self.state.watcher_path_input.value.strip()
-                    if self.state.watcher_path_input is not None
-                    else ""
-                )
-                if watcher_path:
-                    pt_yaml.set_watcher_file_path(watcher_path)
-                pt_yaml.write(str(yaml_file))
-                _LOGGER.info("Powertrain YAML config saved to %s", yaml_file)
-            except Exception:
-                _LOGGER.exception("Failed to write YAML config to %s.", yaml_file)
-
-        if json_path:
-            json_file = Path(json_path)
-            json_file.parent.mkdir(parents=True, exist_ok=True)
-            canvas_state = {
-                "components": nodes_data,
-                "connections": edges_data,
-                "source_ports": source_port_data,
-                "target_ports": target_port_data,
-            }
-            with open(json_file, "w") as f:
-                json.dump(canvas_state, f, indent=2)
-            _LOGGER.info("Canvas state (JSON backup) saved to %s", json_file)
-
-        if self.state.save_button is not None:
-            self.state.save_button.button_type = "success"
-
-    def _load_canvas_state(self, json_path: str):
-        """
-        Restore the canvas from a JSON canvas-state backup file.
-
-        Used only by the ``prompt()`` fallback where the user types a
-        filesystem path manually.
-
-        :param json_path: Path to the JSON file written by a previous save.
-        """
-        json_file = Path(json_path)
-        if not json_file.exists():
-            _LOGGER.error("Load failed – file not found: %s", json_file)
-            return
-
-        try:
-            with open(json_file) as f:
-                canvas_state = json.load(f)
-        except Exception:
-            _LOGGER.exception("Load failed – could not parse JSON from %s", json_file)
-            return
-
-        _LOGGER.info("Loading canvas state from file: %s", json_file)
-        self._restore_canvas_from_dict(canvas_state)
-
-    def _restore_canvas_from_dict(self, canvas_state: dict):
-        """
-        Shared canvas-restoration logic used by both load paths.
-
-        Clears the current canvas entirely, then replays every node, edge, and
-        port that was serialised by a previous save.
-
-        :param canvas_state: Parsed canvas-state dict with keys ``"components"``,
-            ``"connections"``, ``"source_ports"``, and ``"target_ports"``.
-        """
-        # Since the palette buttons are still active while loading, reset is required
-        self.state.selected_node_index = None
-        self.state.selected_component = None
-        self.state.delete_mode = False
-        self.state.pending_connections.clear()
-        self.state.placed_counter.clear()
-        self._clear_node_table()
-        self._clear_temp_edges()
-
-        # Read from canvas_state extract from JSON, defaulting to empty dicts if keys are missing
-        nodes_data = canvas_state.get("components", {})
-        edges_data = canvas_state.get("connections", {})
-        source_data = canvas_state.get("source_ports", {})
-        target_data = canvas_state.get("target_ports", {})
-
-        # Ensure every expected column exists (forward-compatibility with older saves)
-        _node_defaults = {
-            "x": [],
-            "y": [],
-            "url": [],
-            "w": [],
-            "h": [],
-            "name": [],
-            "node_type": [],
-            "icon_type": [],
-            "position": [],
-            "options": [],
-            "n_sources": [],
-            "n_targets": [],
-            "symmetry_name": [],
-            "symmetry_node_index": [],
-        }
-        for col, default in _node_defaults.items():
-            nodes_data.setdefault(col, default)
-
-        # Fill the data sources with the loaded data
-        self.state.placed_nodes_source.data = {
-            key: list(value) for key, value in nodes_data.items()
-        }
-        self.state.edge_source.data = {key: list(value) for key, value in edges_data.items()}
-        self.state.source_port_source.data = {
-            key: list(value) for key, value in source_data.items()
-        }
-        self.state.target_port_source.data = {
-            key: list(value) for key, value in target_data.items()
-        }
-
-        self.state.hover_source.data = {
-            "x": list(nodes_data.get("x", [])),
-            "y": list(nodes_data.get("y", [])),
-            "name": list(nodes_data.get("name", [])),
-            "node_type": list(nodes_data.get("node_type", [])),
-        }
-
-        # Rebuild placed_counter so future placements continue from the right index
-        for name in nodes_data.get("name", []):
-            parts = name.rsplit("_", 1)
-            if len(parts) == 2 and parts[1].isdigit():
-                key = parts[0]
-                self.state.placed_counter[key] = max(
-                    self.state.placed_counter.get(key, 0), int(parts[1])
-                )
-
-        # Redraw the canvas elements based on the restored data sources
-        self._rebuild_all_ports()
-        self._rebuild_edges()
-
-        if self.state.save_button is not None:
-            self.state.save_button.button_type = "success"
-
-        _LOGGER.info(
-            "Canvas restored (%d nodes, %d edges).",
-            len(nodes_data.get("name", [])),
-            len(edges_data.get("xs", [])),
-        )
-
-    def _on_browse_load(self, attr, old, new):
-        """
-        Open a native OS open-file dialog (tkinter) for loading a JSON canvas
-        backup, then restore the canvas from the chosen file.
-
-        Triggered by the hidden ``browse_load_trigger`` TextInput toggling.
-        """
-        path = self._open_tkinter_dialog(
-            filedialog.askopenfilename,
-            title="Load canvas state",
-            filetypes=[("JSON canvas backup", "*.json"), ("All files", "*.*")],
-        )
-        if not path:
-            return
-        self._dismiss_startup_overlay()
-        self._load_canvas_state(json_path=path)
-
-    @staticmethod
-    def _open_tkinter_dialog(func, **kwargs):
-        """
-        Run a tkinter file dialog on the calling thread and return the result.
-
-        :param func: The ``filedialog`` function to call
-            (e.g. ``filedialog.asksaveasfilename``).
-        :param kwargs: Keyword arguments forwarded to *func*.
-        :return: The chosen path string, or ``""`` if the user cancelled.
-        """
-
-        root = tk.Tk()
-        root.withdraw()
-        root.attributes("-topmost", True)
-        try:
-            result = func(parent=root, **kwargs)
-        finally:
-            root.destroy()
-        return result or ""
-
-    # -----------------------------------------------------------------------
-    # Save / load (continued)
-    # -----------------------------------------------------------------------
-
-    def _on_browse_watcher(self, attr, old, new):
-        """
-        Open a native OS save-file dialog (tkinter) to choose the watcher CSV
-        path and write it back into ``state.watcher_path_input``.
-
-        Triggered by the hidden ``browse_watcher_trigger`` TextInput toggling.
-        """
-        path = self._open_tkinter_dialog(
-            filedialog.asksaveasfilename,
-            title="Choose watcher CSV file",
-            defaultextension=".csv",
-            filetypes=[("CSV files", "*.csv"), ("All files", "*.*")],
-        )
-        if path and self.state.watcher_path_input is not None:
-            self.state.watcher_path_input.value = path
-            _LOGGER.info("Watcher CSV path set to: %s", path)
-
-    # -----------------------------------------------------------------------
-    # End session
-    # -----------------------------------------------------------------------
-
-    def _end_session(self):
-        """
-        Stop the Bokeh IO loop, terminating the server session.
-
-        Never called directly from a button ``on_click``; always invoked via
-        :meth:`_on_end_session_trigger` so that the JS gate (which checks the
-        save-button colour) has already run before Python acts.
-        """
-        _LOGGER.info("Ending session and stopping server")
-        IOLoop.current().stop()
-
-    def _on_end_session_trigger(self, attr, old, new):
-        """
-        Fire :meth:`_end_session` in response to the ``end_session_trigger``
-        TextInput toggle.
-
-        JS flips this trigger in exactly two situations:
-
-        * The End Session button is clicked **and** the design is already saved
-          (save button is green).
-        * The **End Anyway** button in the unsaved-exit overlay is clicked.
-
-        By routing through a trigger rather than ``on_click`` we guarantee that
-        Python's stop call only runs when JS has already decided it is safe to
-        do so.
-        """
-        self._end_session()
-
-    def _on_end_session_save_trigger(self, attr, old, new):
-        """
-        Triggered by the ``end_session_save_trigger`` TextInput when the user
-        clicks **Save & Exit** in the unsaved-exit overlay.
-
-        The JS side has already opened the watcher-path overlay for the user
-        to fill in; once they click **Continue to Save** the
-        ``browse_save_trigger`` fires ``_on_browse_save``.  We simply set an
-        internal flag here so that ``_on_browse_save`` knows it should call
-        ``_end_session`` after saving.
-        """
-        _LOGGER.info("Save & Exit requested – will end session after save completes.")
-        self._pending_exit_after_save = True
-
-    def _on_browse_save(self, attr, old, new):
-        """
-        Open native OS save-file dialogs (tkinter) for the YAML config and JSON
-        backup, then write both files.  The watcher file path is read from
-        ``state.watcher_path_input`` which the user filled in the overlay.
-
-        If ``_pending_exit_after_save`` is ``True`` (set by
-        ``_on_end_session_save_trigger``), the session is ended automatically
-        after a successful save.
-
-        Triggered by the hidden ``browse_save_trigger`` TextInput toggling.
-        """
-
-        yaml_path = self._open_tkinter_dialog(
-            filedialog.asksaveasfilename,
-            title="Save YAML powertrain config",
-            defaultextension=".yml",
-            filetypes=[("YAML files", "*.yml *.yaml"), ("All files", "*.*")],
-        )
-        json_path = self._open_tkinter_dialog(
-            filedialog.asksaveasfilename,
-            title="Save JSON canvas backup",
-            defaultextension=".json",
-            filetypes=[("JSON files", "*.json"), ("All files", "*.*")],
-        )
-        if yaml_path or json_path:
-            self._save_canvas_state(yaml_path=yaml_path, json_path=json_path)
-
-        if getattr(self, "_pending_exit_after_save", False):
-            self._pending_exit_after_save = False
-            _LOGGER.info("Save & Exit: save complete – signalling browser then ending session.")
-            # Flip close_window_trigger BEFORE stopping the IOLoop so Bokeh can
-            # push the value change to the browser (which fires window.close() via
-            # js_on_change).  _end_session() calls IOLoop.stop(), after which no
-            # further document updates can be delivered to the client.
-            if self.state.close_window_trigger is not None:
-                self.state.close_window_trigger.value = (
-                    "0" if self.state.close_window_trigger.value == "1" else "1"
-                )
-            # Defer the actual IOLoop stop by one tick so Bokeh has time to flush
-            # the trigger update to the WebSocket before the server shuts down.
-            IOLoop.current().call_later(0.3, self._end_session)
-
-    def _find_nearest_port(self, x: float, y: float) -> dict | None:
-        """
-        Return metadata for the nearest port ball within snap distance, or ``None``.
-
-        :param x: Tap x coordinate in canvas data-units.
-        :param y: Tap y coordinate in canvas data-units.
-
-        :return: A dict with keys ``kind``, ``x``, ``y``, ``color``, ``label``,
-            and ``node_index`` for the nearest port, or ``None``.
-        """
-        snap = PORT_RADIUS * 2.5
-        nearest_port, nearest_port_distance = None, snap
-        for kind, port_source in [
-            ("source", self.state.source_port_source),
-            ("target", self.state.target_port_source),
-        ]:
-            if port_source is None:
-                continue
-            data = port_source.data
-            for i, (port_x, port_y) in enumerate(
-                zip(list(data.get("x", [])), list(data.get("y", [])))
-            ):
-                distance = ((x - port_x) ** 2 + (y - port_y) ** 2) ** 0.5
-                if distance < nearest_port_distance:
-                    nearest_port_distance = distance
-                    nearest_port = {
-                        "kind": kind,
-                        "x": port_x,
-                        "y": port_y,
-                        "color": list(data["color"])[i],
-                        "label": list(data["label"])[i],
-                        "node_index": list(data["node_index"])[i],
-                    }
-        return nearest_port
-
-    def _cancel_pending_connection(self):
-        """Cancel any port selection waiting for a second click."""
-        self.state.pending_port = None
-        if self.state.pending_port_source is not None:
-            self.state.pending_port_source.data = dict(x=[], y=[], color=[])
-
-    def _handle_port_tap(self, port: dict):
-        """
-        Manage the two-click port-connection workflow.
-
-        First click stores the port as pending; second click on a compatible
-        port (opposite kind, same energy colour) draws a permanent edge.
-        Tapping the same port again cancels the pending selection.
-
-        :param port: Port metadata dict with keys ``kind``, ``x``, ``y``,
-            ``color``, ``label``, and ``node_index``.
-        """
-        self._clear_temp_edges()
-        pending = self.state.pending_port
-        if pending is None:
-            self.state.pending_port = port
-            if self.state.pending_port_source is not None:
-                self.state.pending_port_source.data = dict(
-                    x=[port["x"]], y=[port["y"]], color=[port["color"]]
-                )
-            return
-        # Same port tapped again → cancel
-        if (
-            port["kind"] == pending["kind"]
-            and port["node_index"] == pending["node_index"]
-            and port["label"] == pending["label"]
-        ):
-            self._cancel_pending_connection()
-            return
-        # Connect source ↔ target of the same energy type (colour)
-        if port["kind"] != pending["kind"] and port["color"] == pending["color"]:
-            self._add_edge(pending, port)
-            self._rebuild_all_ports()
-        self._cancel_pending_connection()
-
-    def _cursor_to_segment_distance(
-        self, px: float, py: float, x1: float, y1: float, x2: float, y2: float
-    ) -> float:
-        """
-        Return the minimum Euclidean distance from ``(px, py)`` to the segment
-        ``(x1, y1)``–``(x2, y2)``.
-
-        Used by :meth:`_find_nearest_edge` for click-to-delete on edges.
-        """
-        segment_delta_x, segment_delta_y = x2 - x1, y2 - y1
-        if segment_delta_x == 0 and segment_delta_y == 0:
-            return ((px - x1) ** 2 + (py - y1) ** 2) ** 0.5
-
-        projection_parameter = max(
-            0.0,
-            min(
-                1.0,
-                ((px - x1) * segment_delta_x + (py - y1) * segment_delta_y)
-                / (segment_delta_x**2 + segment_delta_y**2),
-            ),
-        )
-        return (
-            (px - (x1 + projection_parameter * segment_delta_x)) ** 2
-            + (py - (y1 + projection_parameter * segment_delta_y)) ** 2
-        ) ** 0.5
-
-    def _find_nearest_edge(self, x: float, y: float, snap: float = 12.0) -> int | None:
-        """
-        Return the index of the edge closest to ``(x, y)`` within *snap* pixels,
-        or ``None`` if no edge is close enough.
-
-        :param x: Tap x coordinate in canvas data-units.
-        :param y: Tap y coordinate in canvas data-units.
-        :param snap: Maximum allowed distance (data-units) for a hit.
-
-        :return: Zero-based row index into ``edge_source``, or ``None``.
-        """
-        if self.state.edge_source is None:
-            return None
-
-        edge_data = self.state.edge_source.data
-        best_index, best_distance = None, snap
-        for i, (xs, ys) in enumerate(zip(edge_data.get("xs", []), edge_data.get("ys", []))):
-            if len(xs) < 2:
-                continue
-            distance = self._cursor_to_segment_distance(x, y, xs[0], ys[0], xs[1], ys[1])
-            if distance < best_distance:
-                best_distance = distance
-                best_index = i
-        return best_index
-
-    def _add_edge(self, starting_port: dict, ending_port: dict):
-        """
-        Append a permanent edge between two ports to ``edge_source``.
-
-        Guards against already-occupied ports and loop connections; rejects
-        the edge silently if either check fails.
-
-        :param starting_port: Metadata dict for the source-side (starting) port.
-        :param ending_port: Metadata dict for the target-side (ending) port.
-        """
-        if self.state.edge_source is None:
-            return
-        edge_data = {key: list(value) for key, value in self.state.edge_source.data.items()}
-
-        for i in range(len(edge_data.get("starting_node_index", []))):
-            starting_port_already_used = (
-                edge_data["starting_node_index"][i] == starting_port["node_index"]
-                and edge_data["starting_port_label"][i] == starting_port["label"]
-                and edge_data["starting_port_kind"][i] == starting_port["kind"]
-            ) or (
-                edge_data["ending_node_index"][i] == starting_port["node_index"]
-                and edge_data["ending_port_label"][i] == starting_port["label"]
-                and edge_data["ending_port_kind"][i] == starting_port["kind"]
-            )
-            ending_port_already_used = (
-                edge_data["starting_node_index"][i] == ending_port["node_index"]
-                and edge_data["starting_port_label"][i] == ending_port["label"]
-                and edge_data["starting_port_kind"][i] == ending_port["kind"]
-            ) or (
-                edge_data["ending_node_index"][i] == ending_port["node_index"]
-                and edge_data["ending_port_label"][i] == ending_port["label"]
-                and edge_data["ending_port_kind"][i] == ending_port["kind"]
-            )
-            if starting_port_already_used or ending_port_already_used:
-                _LOGGER.info("Port already connected; edge rejected.")
-                return
-
-        loop_connection = any(
-            (
-                edge_data["starting_node_index"][i] == ending_port["node_index"]
-                and edge_data["ending_node_index"][i] == starting_port["node_index"]
-            )
-            or (
-                edge_data["starting_node_index"][i] == starting_port["node_index"]
-                and edge_data["ending_node_index"][i] == ending_port["node_index"]
-            )
-            for i in range(len(edge_data.get("starting_node_index", [])))
-        )
-        if loop_connection:
-            _LOGGER.info("Loop connection detected; edge rejected.")
-            return
-
-        edge_data["xs"].append([starting_port["x"], ending_port["x"]])
-        edge_data["ys"].append([starting_port["y"], ending_port["y"]])
-        edge_data["color"].append(starting_port["color"])
-        edge_data["starting_node_index"].append(starting_port["node_index"])
-        edge_data["starting_port_label"].append(starting_port["label"])
-        edge_data["starting_port_kind"].append(starting_port["kind"])
-        edge_data["ending_node_index"].append(ending_port["node_index"])
-        edge_data["ending_port_label"].append(ending_port["label"])
-        edge_data["ending_port_kind"].append(ending_port["kind"])
-        self.state.edge_source.data = edge_data
-
-        if self.state.selected_node_index is not None and (
-            starting_port["node_index"] == self.state.selected_node_index
-            or ending_port["node_index"] == self.state.selected_node_index
-        ):
-            source_data = {
-                key: list(value) for key, value in self.state.source_port_source.data.items()
-            }
-            target_data = {
-                key: list(value) for key, value in self.state.target_port_source.data.items()
-            }
-
-            def _mark_connected(data, port):
-                data["connected"] = [
-                    "True"
-                    if (
-                        data["node_index"][i] == port["node_index"]
-                        and data["label"][i] == port["label"]
-                    )
-                    else data["connected"][i]
-                    for i in range(len(data["node_index"]))
-                ]
-
-            if starting_port["kind"] == "source":
-                _mark_connected(source_data, starting_port)
-                _mark_connected(target_data, ending_port)
-            else:
-                _mark_connected(target_data, starting_port)
-                _mark_connected(source_data, ending_port)
-
-            self.state.source_port_source.data = source_data
-            self.state.target_port_source.data = target_data
-            self._refresh_connections_table(self.state.selected_node_index)
-
-        self._mark_unsaved()
-        _LOGGER.info(
-            "Edge: %s port %s (node %d) ↔ %s port %s (node %d)",
-            starting_port["kind"],
-            starting_port["label"],
-            starting_port["node_index"],
-            ending_port["kind"],
-            ending_port["label"],
-            ending_port["node_index"],
-        )
-
-    def _add_edge_temp(self, starting_port: dict, ending_port: dict):
-        """
-        Draw a dashed preview line between two ports in ``temp_edge_source``.
-
-        The line is **not** registered in ``edge_source``; it becomes permanent
-        only when :meth:`apply_node_configurations` is called.
-        """
-        if self.state.temp_edge_source is None:
-            return
-        temp_edge_data = {
-            key: list(value) for key, value in self.state.temp_edge_source.data.items()
-        }
-        temp_edge_data["xs"].append([starting_port["x"], ending_port["x"]])
-        temp_edge_data["ys"].append([starting_port["y"], ending_port["y"]])
-        temp_edge_data["color"].append(starting_port["color"])
-        self.state.temp_edge_source.data = temp_edge_data
-        _LOGGER.info(
-            "Temp edge: %s:%s → %s:%s",
-            starting_port["kind"],
-            starting_port["label"],
-            ending_port["kind"],
-            ending_port["label"],
-        )
-
-    def _clear_temp_edges(self):
-        """Wipe all dashed preview edges and the pending-connection list."""
-        self.state.pending_connections.clear()
-        self._clear_temp_edge_visuals()
-
-    def _clear_temp_edge_visuals(self):
-        """Clear only the canvas dashed lines, leave pending_connections intact."""
-        if self.state.temp_edge_source is not None:
-            self.state.temp_edge_source.data = dict(xs=[], ys=[], color=[])
-
-    def _rebuild_edges(self):
-        """
-        Recompute xs/ys in edge_source from stored port-identity columns.
-
-        Called at the end of :meth:`_rebuild_all_ports` so edges track port
-        positions after port-count spinner changes.  Edges whose ports no
-        longer exist (e.g. port count reduced) are silently dropped.
-        """
-        if self.state.edge_source is None:
-            return
-        edge_data = {key: list(value) for key, value in self.state.edge_source.data.items()}
-        if not edge_data.get("starting_node_index"):
-            return
-
-        position: dict = {}
-        for kind, data_source in [
-            ("source", self.state.source_port_source),
-            ("target", self.state.target_port_source),
-        ]:
-            if data_source is None:
-                continue
-            dataset = data_source.data
-            for i, (px, py) in enumerate(
-                zip(list(dataset.get("x", [])), list(dataset.get("y", [])))
-            ):
-                key = (kind, list(dataset["node_index"])[i], list(dataset["label"])[i])
-                position[key] = (px, py)
-
-        new_xs, new_ys, valid = [], [], []
-        for i in range(len(edge_data["starting_node_index"])):
-            starting_port_key = (
-                edge_data["starting_port_kind"][i],
-                edge_data["starting_node_index"][i],
-                edge_data["starting_port_label"][i],
-            )
-            ending_port_key = (
-                edge_data["ending_port_kind"][i],
-                edge_data["ending_node_index"][i],
-                edge_data["ending_port_label"][i],
-            )
-            if starting_port_key in position and ending_port_key in position:
-                starting_x, starting_y = position[starting_port_key]
-                ending_x, ending_y = position[ending_port_key]
-                new_xs.append([starting_x, ending_x])
-                new_ys.append([starting_y, ending_y])
-                valid.append(i)
-
-        new_edge_data = {k: [edge_data[k][j] for j in valid] for k in edge_data}
-        new_edge_data["xs"] = new_xs
-        new_edge_data["ys"] = new_ys
-        self.state.edge_source.data = new_edge_data
-
-    # -----------------------------------------------------------------------
-    # Port management
-    # -----------------------------------------------------------------------
-
-    def _rebuild_all_ports(self):
-        """
-        Recompute every port ball position from the current placed-nodes data
-        and push the result into ``source_port_source`` / ``target_port_source``.
-
-        Called whenever nodes are added, moved, deleted, or their port counts
-        are changed via the spinner widgets.
-        """
-        if self.state.source_port_source is None or self.state.target_port_source is None:
-            return
-
-        selected_node_fill_alpha, selected_node_line_alpha = 0.3, 0.5
-
-        placed_nodes_data = self.state.placed_nodes_source.data
-        xs = list(placed_nodes_data.get("x", []))
-        ys = list(placed_nodes_data.get("y", []))
-        icon_types = list(placed_nodes_data.get("icon_type", []))
-        node_types = list(placed_nodes_data.get("node_type", []))
-        node_names = list(placed_nodes_data.get("name", []))
-        n_sources_list = list(placed_nodes_data.get("n_sources", []))
-        n_targets_list = list(placed_nodes_data.get("n_targets", []))
-
-        source_x, source_y, source_color, source_label = [], [], [], []
-        source_node_index, source_node_name, source_node_type = [], [], []
-        source_fill_alpha, source_line_alpha = [], []
-
-        target_x, target_y, target_color, target_label = [], [], [], []
-        target_node_index, target_node_name, target_node_type = [], [], []
-        target_fill_alpha, target_line_alpha = [], []
-
-        connected_ports: set = set()
-        if self.state.edge_source is not None:
-            edge_data = self.state.edge_source.data
-            for i in range(len(edge_data.get("starting_node_index", []))):
-                connected_ports.add(
-                    (
-                        edge_data["starting_port_kind"][i],
-                        edge_data["starting_node_index"][i],
-                        edge_data["starting_port_label"][i],
-                    )
-                )
-                connected_ports.add(
-                    (
-                        edge_data["ending_port_kind"][i],
-                        edge_data["ending_node_index"][i],
-                        edge_data["ending_port_label"][i],
-                    )
-                )
-
-        has_selected_node = self.state.selected_node_index is not None
-        selected_node_index = self.state.selected_node_index
-
-        for i, (cx, cy, icon_type, node_type, node_name) in enumerate(
-            zip(xs, ys, icon_types, node_types, node_names)
-        ):
-            n_source = (
-                int(n_sources_list[i])
-                if i < len(n_sources_list)
-                else self.state.default_source_count.get(node_type, 0)
-            )
-            n_target = (
-                int(n_targets_list[i])
-                if i < len(n_targets_list)
-                else self.state.default_target_count.get(node_type, 0)
-            )
-
-            cfg = ICONS_CONFIG.get(icon_type, {})
-            raw_source_color = cfg.get("source_color") or DEFAULT_COLOR
-            raw_target_color = cfg.get("target_color") or DEFAULT_COLOR
-
-            ports = compute_ports(cx, cy, NODE_RADIUS, PORT_RADIUS, n_source, n_target)
-            is_selected = has_selected_node and i == selected_node_index
-            fill_alpha = selected_node_fill_alpha if is_selected else 0.0
-            line_alpha = selected_node_line_alpha if is_selected else 0.0
-
-            for port in ports["outputs"]:
-                source_x.append(port["x"])
-                source_y.append(port["y"])
-                source_color.append(raw_source_color)
-                source_label.append(str(port["index"] + 1))
-                source_node_index.append(i)
-                source_node_name.append([node_name])
-                source_node_type.append([node_type])
-                source_fill_alpha.append(fill_alpha)
-                source_line_alpha.append(line_alpha)
-
-            for port in ports["inputs"]:
-                target_x.append(port["x"])
-                target_y.append(port["y"])
-                target_color.append(raw_target_color)
-                target_label.append(str(port["index"] + 1))
-                target_node_index.append(i)
-                target_node_name.append([node_name])
-                target_node_type.append([node_type])
-                target_fill_alpha.append(fill_alpha)
-                target_line_alpha.append(line_alpha)
-
-        self.state.source_port_source.data = dict(
-            x=source_x,
-            y=source_y,
-            color=source_color,
-            label=source_label,
-            node_index=source_node_index,
-            node_name=source_node_name,
-            node_type=source_node_type,
-            fill_alpha=source_fill_alpha,
-            line_alpha=source_line_alpha,
-            kind=["source"] * len(source_x),
-            connected=[
-                "True"
-                if ("source", source_node_index[i], source_label[i]) in connected_ports
-                else "False"
-                for i in range(len(source_x))
-            ],
-        )
-        self.state.target_port_source.data = dict(
-            x=target_x,
-            y=target_y,
-            color=target_color,
-            label=target_label,
-            node_index=target_node_index,
-            node_name=target_node_name,
-            node_type=target_node_type,
-            fill_alpha=target_fill_alpha,
-            line_alpha=target_line_alpha,
-            kind=["target"] * len(target_x),
-            connected=[
-                "True"
-                if ("target", target_node_index[i], target_label[i]) in connected_ports
-                else "False"
-                for i in range(len(target_x))
-            ],
-        )
-        self._rebuild_edges()
-
-        if self.state.selected_node_index is not None:
-            self._refresh_connections_table(self.state.selected_node_index)
-
-    def _best_possible_node(self, x: float, y: float):
-        """
-        Find the nearest placed node to canvas coordinates ``(x, y)``.
-
-        :param x: Tap x coordinate in canvas data-units.
-        :param y: Tap y coordinate in canvas data-units.
-
-        :return: ``(best_index, shortest_distance, nodes_data)`` where ``best_index`` is the
-            zero-based index of the closest node (``None`` if nothing is within snap
-            distance), ``shortest_distance`` is the Euclidean distance, and ``nodes_data``
-            is the raw ``placed_nodes_source.data`` dict.
-        """
-        nodes_data = self.state.placed_nodes_source.data
-        xs = list(nodes_data.get("x", []))
-        ys = list(nodes_data.get("y", []))
-        if not xs:
-            return None, None, nodes_data
-
-        snap = self.icon_size
-        best_index = None
-        shortest_distance = float("inf")
-        for node_position_index, (node_x, node_y) in enumerate(zip(xs, ys)):
-            distance_to_tap = ((x - node_x) ** 2 + (y - node_y) ** 2) ** 0.5
-            # Only consider nodes within snap distance and track the closest one
-            if distance_to_tap < snap and distance_to_tap < shortest_distance:
-                shortest_distance = distance_to_tap
-                best_index = node_position_index
-
-        return best_index, shortest_distance, nodes_data
-
-    # -----------------------------------------------------------------------
-    # Config panel helpers
-    # -----------------------------------------------------------------------
-
-    def _populate_node_table(self, index: int):
-        """
-        Fill the config panel widgets with the stored values of node *index*.
-
-        :param index: Zero-based index into ``placed_nodes_source`` data arrays.
-        """
-        placed_node_data = self.state.placed_nodes_source.data
-        node_name = list(placed_node_data.get("name", []))[index]
-        node_type = list(placed_node_data.get("node_type", []))[index]
-        icon_key = list(placed_node_data.get("icon_type", []))[index]
-        position = (
-            list(placed_node_data.get("position", []))[index]
-            if placed_node_data.get("position")
-            else _EMPTY
-        )
-        saved_options_json = (
-            list(placed_node_data.get("options", []))[index]
-            if placed_node_data.get("options")
-            else "{}"
-        )
-
-        self.state.options_table.visible = saved_options_json != "{}"
-        self.state.name_input.value = node_name
-
-        choices = self.state.component_type_to_icon.get(icon_key, icon_key)
-        self.state.type_select.options = choices
-        self.state.type_select.value = node_type if node_type in choices else choices[0]
-
-        pos_choices = self.state.possible_position.get(node_type, [])
-        self.state.position_select.options = pos_choices
-        if pos_choices:
-            self.state.position_select.value = (
-                position if position in pos_choices else pos_choices[0]
-            )
-        else:
-            self.state.position_select.value = ""
-
-        saved_options: dict = {}
-        try:
-            saved_options = json.loads(saved_options_json) if saved_options_json else {}
-        except (json.JSONDecodeError, TypeError):
-            pass
-
-        self._refresh_options_table(node_type, saved_options)
-
-        default_source_port_count = self.state.default_source_count.get(node_type, 0)
-        default_target_port_count = self.state.default_target_count.get(node_type, 0)
-        source_ports_are_editable = default_source_port_count == 3
-        target_ports_are_editable = default_target_port_count == 3
-
-        current_source_port_count = (
-            list(placed_node_data.get("n_sources", []))[index]
-            if placed_node_data.get("n_sources")
-            else default_source_port_count
-        )
-        current_target_port_count = (
-            list(placed_node_data.get("n_targets", []))[index]
-            if placed_node_data.get("n_targets")
-            else default_target_port_count
-        )
-
-        if self.state.source_count_spinner is not None:
-            self.state.source_count_spinner.visible = source_ports_are_editable
-            if source_ports_are_editable:
-                self.state.source_count_spinner.value = int(current_source_port_count)
-        if self.state.target_count_spinner is not None:
-            self.state.target_count_spinner.visible = target_ports_are_editable
-            if target_ports_are_editable:
-                self.state.target_count_spinner.value = int(current_target_port_count)
-        if self.state.port_count_section is not None:
-            self.state.port_count_section.visible = (
-                source_ports_are_editable or target_ports_are_editable
-            )
-
-        if self.state.selected_node_overlay_source is not None:
-            self.state.selected_node_overlay_source.data = dict(
-                x=[list(placed_node_data["x"])[index]],
-                y=[list(placed_node_data["y"])[index]],
-            )
-        self._rebuild_all_ports()
-        self._refresh_connections_table(index)
-        self._refresh_symmetry_select(index, icon_key)
-
-    @staticmethod
-    def _option_values_to_strings(value) -> str:
-        """Convert an option value (bool, int, float, or str) to its display string."""
-        if value is True:
-            return "True"
-        if value is False:
-            return "False"
-        return str(value)
-
-    @staticmethod
-    def _strings_to_option_values(string: str):
-        """Parse a display string back to a typed Python value (bool, int, float, or str)."""
-        if string == "True":
-            return True
-        if string == "False":
-            return False
-        try:
-            return int(string)
-        except (ValueError, TypeError):
-            pass
-        try:
-            return float(string)
-        except (ValueError, TypeError):
-            pass
-        return string
-
-    def _refresh_options_table(self, node_type: str, overrides: dict = None):
-        """
-        Rebuild the Options section of the config panel for a given node type.
-
-        Creates one row per option (a disabled TextInput label + a Select for
-        the value) and wires a shared callback so that ``options_source`` stays
-        in sync with the user's current selections.
-
-        :param node_type: Component type key used to look up ``possible_options``.
-        :param overrides: Optional ``{option_name: current_value}`` mapping used
-            to pre-populate widgets with previously saved values.
-        """
-        if overrides is None:
-            overrides = {}
-        options_definition = self.state.possible_options.get(node_type, {})
-        option_names = list(options_definition.keys())
-        opt_values = []
-        new_rows = []
-        value_selects = []
-
-        for option_name, value_list in options_definition.items():
-            current_values = (
-                overrides[option_name]
-                if option_name in overrides
-                else (value_list[0] if value_list else _EMPTY)
-            )
-            current_stringing = self._option_values_to_strings(current_values)
-
-            label_input = bkmodel.TextInput(
-                value=option_name,
-                width=180,
-                disabled=True,
-                styles={"color": "white", "font-size": "14px"},
-            )
-            choices = (
-                [self._option_values_to_strings(choice) for choice in value_list]
-                if value_list
-                else [current_stringing]
-            )
-            if current_stringing not in choices:
-                choices = [current_stringing] + choices
-
-            value_select = bkmodel.Select(
-                value=current_stringing,
-                options=choices,
-                width=180,
-                styles={"color": "white", "font-size": "14px"},
-            )
-            new_rows.append(row(label_input, value_select, spacing=4))
-            opt_values.append(current_stringing)
-            value_selects.append(value_select)
-
-        def _make_sync_callback(selects, names):
-            def _on_change(attr, old, new):
-                self.state.options_source.data = dict(
-                    options=names,
-                    value=[selected.value for selected in selects],
-                )
-
-            return _on_change
-
-        sync_callback = _make_sync_callback(value_selects, option_names)
-        for value_selected in value_selects:
-            value_selected.on_change("value", sync_callback)
-
-        if self.state.options_rows_column is not None:
-            self.state.options_rows_column.children = new_rows
-        self.state.options_source.data = dict(options=option_names, value=opt_values)
-
-    def _refresh_symmetry_select(self, current_node_index: int, icon_type: str):
-        """
-        Populate ``symmetry_select`` with the names of all placed nodes that
-        share *icon_type* with the currently selected node, excluding itself.
-
-        :param current_node_index: Index of the node being edited.
-        :param icon_type: The ``icon_type`` key of the selected node.
-        """
-        if self.state.symmetry_select is None:
-            return
-
-        placed_nodes_data = self.state.placed_nodes_source.data
-        names = list(placed_nodes_data.get("name", []))
-        icon_types = list(placed_nodes_data.get("icon_type", []))
-        saved_symmetry = list(placed_nodes_data.get("symmetry_name", []))
-
-        symmetry_candidates = [
-            names[i]
-            for i in range(len(names))
-            if i != current_node_index and icon_types[i] == icon_type
-        ]
-        choices = [_EMPTY] + symmetry_candidates
-        self.state.symmetry_select.options = choices
-
-        current_symmetry_component = (
-            saved_symmetry[current_node_index]
-            if current_node_index < len(saved_symmetry)
-            else _EMPTY
-        )
-        self.state.symmetry_select.value = (
-            current_symmetry_component if current_symmetry_component in choices else _EMPTY
-        )
-
-    def _clear_node_table(self):
-        """
-        Reset all node configuration panel inputs to their empty defaults and
-        hide the panel.
-        """
-        self.state.name_input.value = _EMPTY
-        self.state.type_select.options = []
-        self.state.type_select.value = _EMPTY
-        self.state.position_select.options = []
-        self.state.position_select.value = _EMPTY
-        self.state.options_source.data = dict(options=[], value=[])
-        if self.state.port_count_section is not None:
-            self.state.port_count_section.visible = False
-            if self.state.source_count_spinner is not None:
-                self.state.source_count_spinner.visible = False
-            if self.state.target_count_spinner is not None:
-                self.state.target_count_spinner.visible = False
-        if self.state.table_panel is not None:
-            self.state.table_panel.visible = False
-        if self.state.selected_node_overlay_source is not None:
-            self.state.selected_node_overlay_source.data = dict(x=[], y=[])
-        if self.state.connections_source is not None:
-            self.state.connections_source.data = dict(my_port=[], connected_to=[], edge_idx=[])
-        if self.state.connections_rows_column is not None:
-            self.state.connections_rows_column.children = []
-        if self.state.symmetry_select is not None:
-            self.state.symmetry_select.options = [_EMPTY]
-            self.state.symmetry_select.value = _EMPTY
-        self._rebuild_all_ports()
-
-    def _refresh_connections_table(self, node_index: int):
-        """
-        Populate the Connections panel for node *node_index* using dynamic rows.
-
-        For every source port of the selected node a row is created with a
-        disabled label and a Select showing compatible unconnected target ports.
-        A symmetric row is created for every target port.  Selecting a new
-        value immediately draws (or removes) the corresponding preview edge.
-
-        :param node_index: Zero-based index of the node to show connections for.
-        """
-        if (
-            self.state.connections_rows_column is None
-            or self.state.edge_source is None
-            or self.state.source_port_source is None
-            or self.state.target_port_source is None
-        ):
-            return
-
-        source_data = {
-            key: list(value) for key, value in self.state.source_port_source.data.items()
-        }
-        target_data = {
-            key: list(value) for key, value in self.state.target_port_source.data.items()
-        }
-        node_data = {key: list(value) for key, value in self.state.placed_nodes_source.data.items()}
-        edge_data = {key: list(value) for key, value in self.state.edge_source.data.items()}
-
-        # Build lookup: port label → (far_node_index, far_label) for every edge
-        # that touches node_index.
-        connected_source: dict[str, tuple[int, str]] = {}
-        connected_target: dict[str, tuple[int, str]] = {}
-
-        for i in range(len(edge_data.get("starting_node_index", []))):
-            starting_node_position = edge_data["starting_node_index"][i]
-            ending_node_position = edge_data["ending_node_index"][i]
-            starting_port_label_value = edge_data["starting_port_label"][i]
-            ending_port_label_value = edge_data["ending_port_label"][i]
-            starting_port_kind_value = edge_data["starting_port_kind"][i]
-            ending_port_kind_value = edge_data["ending_port_kind"][i]
-
-            if starting_node_position == node_index and starting_port_kind_value == "source":
-                connected_source[starting_port_label_value] = (
-                    ending_node_position,
-                    ending_port_label_value,
-                )
-            elif ending_node_position == node_index and ending_port_kind_value == "source":
-                connected_source[ending_port_label_value] = (
-                    starting_node_position,
-                    starting_port_label_value,
-                )
-            if starting_node_position == node_index and starting_port_kind_value == "target":
-                connected_target[starting_port_label_value] = (
-                    ending_node_position,
-                    ending_port_label_value,
-                )
-            elif ending_node_position == node_index and ending_port_kind_value == "target":
-                connected_target[ending_port_label_value] = (
-                    starting_node_position,
-                    starting_port_label_value,
-                )
-
-        def _far_string(far_node_index: int, far_label: str, far_kind: str) -> str:
-            name = (
-                node_data["name"][far_node_index]
-                if far_node_index < len(node_data.get("name", []))
-                else f"node_{far_node_index}"
-            )
-            return f"{name} ({far_kind}:{far_label})"
-
-        def _parse_choice(choice_string: str, candidates: list[tuple]) -> tuple | None:
-            for node_index, label, kind, display_string in candidates:
-                if display_string == choice_string:
-                    return node_index, label, kind
-            return None
-
-        selected_source_color = ICONS_CONFIG.get(node_data["icon_type"][node_index], {}).get(
-            "source_color", ""
-        )
-        free_targets: list[tuple[int, str, str, str]] = []
-        for j in range(len(target_data.get("node_index", []))):
-            if (
-                target_data["connected"][j] == "False"
-                and target_data["node_index"][j] != node_index
-                and target_data["color"][j] == selected_source_color
-            ):
-                free_targets.append(
-                    (
-                        target_data["node_index"][j],
-                        target_data["label"][j],
-                        "target",
-                        _far_string(
-                            target_data["node_index"][j], target_data["label"][j], "target"
-                        ),
-                    )
-                )
-
-        _selected_target_color = ICONS_CONFIG.get(node_data["icon_type"][node_index], {}).get(
-            "target_color", ""
-        )
-        free_sources: list[tuple[int, str, str, str]] = []
-        for j in range(len(source_data.get("node_index", []))):
-            if (
-                source_data["connected"][j] == "False"
-                and source_data["node_index"][j] != node_index
-                and source_data["color"][j] == _selected_target_color
-            ):
-                free_sources.append(
-                    (
-                        source_data["node_index"][j],
-                        source_data["label"][j],
-                        "source",
-                        _far_string(
-                            source_data["node_index"][j], source_data["label"][j], "source"
-                        ),
-                    )
-                )
-
-        new_rows = []
-
-        # ── Source-port rows ──────────────────────────────────────────────────
-        for i in range(len(source_data.get("node_index", []))):
-            if source_data["node_index"][i] != node_index:
-                continue
-
-            source_label = source_data["label"][i]
-            is_connected = source_data["connected"][i] == "True"
-
-            label_input = bkmodel.TextInput(
-                value=f"source:{source_label}",
-                width=180,
-                disabled=True,
-                styles={"color": "white", "font-size": "14px"},
-            )
-
-            candidates = list(free_targets)
-            current_string = _EMPTY
-            if is_connected and source_label in connected_source:
-                far_node_i, far_label = connected_source[source_label]
-                current_string = _far_string(far_node_i, far_label, "target")
-                candidates = [
-                    candidate
-                    for candidate in candidates
-                    if not (candidate[0] == far_node_i and candidate[1] == far_label)
-                ] + [
-                    (far_node_i, far_label, "target", _far_string(far_node_i, far_label, "target"))
-                ]
-
-            choices = [_EMPTY] + [
-                candidate[3] for candidate in candidates if candidate[3] != current_string
-            ]
-            if current_string and current_string not in choices:
-                choices = [current_string] + choices
-
-            value_select = bkmodel.Select(
-                value=current_string if current_string else _EMPTY,
-                options=choices,
-                width=180,
-                styles={"color": "white", "font-size": "12px"},
-            )
-
-            def _make_source_callback(
-                _source_label=source_label,
-                _source_node_idx=node_index,
-                _source_color=source_data["color"][i],
-                _source_x=source_data["x"][i],
-                _source_y=source_data["y"][i],
-                _candidates=candidates,
-            ):
-                def _on_change(attr, old, new_value):
-                    self.state.pending_connections = [
-                        (starting_port, ending_port)
-                        for starting_port, ending_port in self.state.pending_connections
-                        if not (
-                            starting_port["kind"] == "source"
-                            and starting_port["node_index"] == _source_node_idx
-                            and starting_port["label"] == _source_label
-                        )
-                    ]
-                    self._clear_temp_edge_visuals()
-                    for starting_port, ending_port in list(self.state.pending_connections):
-                        self._add_edge_temp(starting_port, ending_port)
-
-                    if new_value == _EMPTY:
-                        if self.state.edge_source is not None:
-                            edge_data = {
-                                key: list(value)
-                                for key, value in self.state.edge_source.data.items()
-                            }
-                            keep = [
-                                i
-                                for i in range(len(edge_data.get("starting_node_index", [])))
-                                if not (
-                                    (
-                                        edge_data["starting_node_index"][i] == _source_node_idx
-                                        and edge_data["starting_port_label"][i] == _source_label
-                                        and edge_data["starting_port_kind"][i] == "source"
-                                    )
-                                    or (
-                                        edge_data["ending_node_index"][i] == _source_node_idx
-                                        and edge_data["ending_port_label"][i] == _source_label
-                                        and edge_data["ending_port_kind"][i] == "source"
-                                    )
-                                )
-                            ]
-                            self.state.edge_source.data = {
-                                k: [edge_data[k][j] for j in keep] for k in edge_data
-                            }
-                            self._rebuild_all_ports()
-                            self._refresh_connections_table(_source_node_idx)
-                        return
-
-                    parsed = _parse_choice(new_value, _candidates)
-                    if parsed is None:
-                        return
-                    target_node_i, target_label, _ = parsed
-
-                    target_data = self.state.target_port_source.data
-                    target_x, target_y, target_color = None, None, None
-                    for j in range(len(target_data.get("node_index", []))):
-                        if (
-                            target_data["node_index"][j] == target_node_i
-                            and target_data["label"][j] == target_label
-                        ):
-                            target_x = target_data["x"][j]
-                            target_y = target_data["y"][j]
-                            target_color = target_data["color"][j]
-                            break
-                    if target_x is None:
-                        return
-
-                    live_source_x, live_source_y = _source_x, _source_y
-                    source_data_live = self.state.source_port_source.data
-                    for j in range(len(source_data_live.get("node_index", []))):
-                        if (
-                            source_data_live["node_index"][j] == _source_node_idx
-                            and source_data_live["label"][j] == _source_label
-                        ):
-                            live_source_x = source_data_live["x"][j]
-                            live_source_y = source_data_live["y"][j]
-                            break
-
-                    starting_port = {
-                        "kind": "source",
-                        "node_index": _source_node_idx,
-                        "label": _source_label,
-                        "x": live_source_x,
-                        "y": live_source_y,
-                        "color": _source_color,
-                    }
-                    ending_port = {
-                        "kind": "target",
-                        "node_index": target_node_i,
-                        "label": target_label,
-                        "x": target_x,
-                        "y": target_y,
-                        "color": target_color,
-                    }
-                    self.state.pending_connections.append((starting_port, ending_port))
-                    self._add_edge_temp(starting_port, ending_port)
-
-                return _on_change
-
-            value_select.on_change(
-                "value",
-                _make_source_callback(
-                    _source_label=source_label,
-                    _source_node_idx=node_index,
-                    _source_color=source_data["color"][i],
-                    _source_x=source_data["x"][i],
-                    _source_y=source_data["y"][i],
-                    _candidates=candidates,
-                ),
-            )
-            new_rows.append(row(label_input, value_select, spacing=4))
-
-        # ── Target-port rows ──────────────────────────────────────────────────
-        for i in range(len(target_data.get("node_index", []))):
-            if target_data["node_index"][i] != node_index:
-                continue
-
-            target_label = target_data["label"][i]
-            is_connected = target_data["connected"][i] == "True"
-
-            label_input = bkmodel.TextInput(
-                value=f"target:{target_label}",
-                width=180,
-                disabled=True,
-                styles={"color": "white", "font-size": "14px"},
-            )
-
-            candidates = list(free_sources)
-            current_string = _EMPTY
-            if is_connected and target_label in connected_target:
-                far_node_i, far_label = connected_target[target_label]
-                current_string = _far_string(far_node_i, far_label, "source")
-                candidates = [
-                    candidate
-                    for candidate in candidates
-                    if not (candidate[0] == far_node_i and candidate[1] == far_label)
-                ] + [
-                    (far_node_i, far_label, "source", _far_string(far_node_i, far_label, "source"))
-                ]
-
-            choices = [_EMPTY] + [
-                candidate[3] for candidate in candidates if candidate[3] != current_string
-            ]
-            if current_string and current_string not in choices:
-                choices = [current_string] + choices
-
-            value_select = bkmodel.Select(
-                value=current_string if current_string else _EMPTY,
-                options=choices,
-                width=180,
-                styles={"color": "white", "font-size": "12px"},
-            )
-
-            def _make_target_callback(
-                _target_label=target_label,
-                _target_node_idx=node_index,
-                _target_color=target_data["color"][i],
-                _target_x=target_data["x"][i],
-                _target_y=target_data["y"][i],
-                _candidates=candidates,
-            ):
-                def _on_change(attr, old, new_value):
-                    self.state.pending_connections = [
-                        (starting_port, ending_port)
-                        for starting_port, ending_port in self.state.pending_connections
-                        if not (
-                            ending_port["kind"] == "target"
-                            and ending_port["node_index"] == _target_node_idx
-                            and ending_port["label"] == _target_label
-                        )
-                    ]
-                    self._clear_temp_edge_visuals()
-                    for starting_port, ending_port in list(self.state.pending_connections):
-                        self._add_edge_temp(starting_port, ending_port)
-
-                    if new_value == _EMPTY:
-                        if self.state.edge_source is not None:
-                            edge_data = {
-                                key: list(value)
-                                for key, value in self.state.edge_source.data.items()
-                            }
-                            keep = [
-                                i
-                                for i in range(len(edge_data.get("starting_node_index", [])))
-                                if not (
-                                    (
-                                        edge_data["starting_node_index"][i] == _target_node_idx
-                                        and edge_data["starting_port_label"][i] == _target_label
-                                        and edge_data["starting_port_kind"][i] == "target"
-                                    )
-                                    or (
-                                        edge_data["ending_node_index"][i] == _target_node_idx
-                                        and edge_data["ending_port_label"][i] == _target_label
-                                        and edge_data["ending_port_kind"][i] == "target"
-                                    )
-                                )
-                            ]
-                            self.state.edge_source.data = {
-                                k: [edge_data[k][j] for j in keep] for k in edge_data
-                            }
-                            self._rebuild_all_ports()
-                            self._refresh_connections_table(_target_node_idx)
-                        return
-
-                    parsed = _parse_choice(new_value, _candidates)
-                    if parsed is None:
-                        return
-                    source_node_i, source_label, _ = parsed
-
-                    source_data = self.state.source_port_source.data
-                    source_x, source_y, source_color = None, None, None
-                    for j in range(len(source_data.get("node_index", []))):
-                        if (
-                            source_data["node_index"][j] == source_node_i
-                            and source_data["label"][j] == source_label
-                        ):
-                            source_x = source_data["x"][j]
-                            source_y = source_data["y"][j]
-                            source_color = source_data["color"][j]
-                            break
-                    if source_x is None:
-                        return
-
-                    live_target_x, live_target_y = _target_x, _target_y
-                    target_data_live = self.state.target_port_source.data
-                    for j in range(len(target_data_live.get("node_index", []))):
-                        if (
-                            target_data_live["node_index"][j] == _target_node_idx
-                            and target_data_live["label"][j] == _target_label
-                        ):
-                            live_target_x = target_data_live["x"][j]
-                            live_target_y = target_data_live["y"][j]
-                            break
-
-                    starting_port = {
-                        "kind": "source",
-                        "node_index": source_node_i,
-                        "label": source_label,
-                        "x": source_x,
-                        "y": source_y,
-                        "color": source_color,
-                    }
-                    ending_port = {
-                        "kind": "target",
-                        "node_index": _target_node_idx,
-                        "label": _target_label,
-                        "x": live_target_x,
-                        "y": live_target_y,
-                        "color": _target_color,
-                    }
-                    self.state.pending_connections.append((starting_port, ending_port))
-                    self._add_edge_temp(starting_port, ending_port)
-
-                return _on_change
-
-            value_select.on_change(
-                "value",
-                _make_target_callback(
-                    _target_label=target_label,
-                    _target_node_idx=node_index,
-                    _target_color=target_data["color"][i],
-                    _target_x=target_data["x"][i],
-                    _target_y=target_data["y"][i],
-                    _candidates=candidates,
-                ),
-            )
-            new_rows.append(row(label_input, value_select, spacing=4))
-
-        self.state.connections_rows_column.children = new_rows
-
-        # Keep connections_source in sync
-        if self.state.connections_source is not None:
-            my_port, connected_to, edge_index = [], [], []
-            for i in range(len(edge_data.get("starting_node_index", []))):
-                starting_node_position = edge_data["starting_node_index"][i]
-                ending_node_position = edge_data["ending_node_index"][i]
-                starting_port_label_value = edge_data["starting_port_label"][i]
-                ending_port_label_value = edge_data["ending_port_label"][i]
-                starting_port_kind_value = edge_data["starting_port_kind"][i]
-                ending_port_kind_value = edge_data["ending_port_kind"][i]
-
-                if starting_node_position == node_index:
-                    far_name = (
-                        node_data["name"][ending_node_position]
-                        if ending_node_position < len(node_data.get("name", []))
-                        else f"node_{ending_node_position}"
-                    )
-                    my_port.append(f"{starting_port_kind_value}:{starting_port_label_value}")
-                    connected_to.append(
-                        f"{far_name} ({ending_port_kind_value}:{ending_port_label_value})"
-                    )
-                    edge_index.append(i)
-                elif ending_node_position == node_index:
-                    far_name = (
-                        node_data["name"][starting_node_position]
-                        if starting_node_position < len(node_data.get("name", []))
-                        else f"node_{starting_node_position}"
-                    )
-                    my_port.append(f"{ending_port_kind_value}:{ending_port_label_value}")
-                    connected_to.append(
-                        f"{far_name} ({starting_port_kind_value}:{starting_port_label_value})"
-                    )
-                    edge_index.append(i)
-
-            self.state.connections_source.data = dict(
-                my_port=my_port,
-                connected_to=connected_to,
-                edge_idx=edge_index,
-            )
-
-    def _delete_selected_connection(self):
-        """
-        Delete the edge(s) currently selected in the Connections DataTable.
-
-        Reads the highlighted row indices from ``connections_source.selected``,
-        resolves them to actual edge indices, removes those rows from
-        ``edge_source``, and refreshes the Connections panel.
-        """
-        if self.state.connections_source is None or self.state.edge_source is None:
-            return
-        selected = list(self.state.connections_source.selected.indices)
-        if not selected:
-            return
-        edge_index_list = list(self.state.connections_source.data.get("edge_idx", []))
-        to_delete = {edge_index_list[i] for i in selected if i < len(edge_index_list)}
-        edge_data = {key: list(value) for key, value in self.state.edge_source.data.items()}
-        keep = [i for i in range(len(edge_data.get("xs", []))) if i not in to_delete]
-        self.state.edge_source.data = {k: [edge_data[k][j] for j in keep] for k in edge_data}
-        self._mark_unsaved()
-        _LOGGER.info("Deleted connection(s) at edge indices %s", sorted(to_delete))
-        if self.state.selected_node_index is not None:
-            self._refresh_connections_table(self.state.selected_node_index)
 
     # -----------------------------------------------------------------------
     # Canvas tap handler
@@ -1881,11 +164,17 @@ class PlacementHandler:
         * **Component selected** – place a new icon at the tap coordinates.
         * **Neither** – select or deselect the nearest existing node.
 
+        An undo snapshot is pushed **before** every action that mutates the
+        canvas (node placement, node deletion, edge deletion) so the user can
+        reverse each step individually.
+
         :param event: Bokeh ``Tap`` event carrying ``x`` and ``y`` coordinates.
         """
         x, y = event.x, event.y
 
-        # Port connection has highest priority in idle mode
+        # Port connection has highest priority in idle mode.
+        # Undo for port connections is pushed inside _handle_port_tap, just
+        # before the permanent edge is committed.
         if not self.state.delete_mode and self.state.selected_component is None:
             nearest_port = self._find_nearest_port(x, y)
             if nearest_port is not None:
@@ -1896,6 +185,9 @@ class PlacementHandler:
             best_node_index, best_node_distance, nodes_data = self._best_possible_node(x, y)
 
             if best_node_index is not None:
+                # Capture state before deleting the node.
+                self._push_undo()
+
                 new_data = {key: list(values) for key, values in nodes_data.items()}
                 for column in new_data:
                     new_data[column].pop(best_node_index)
@@ -1952,6 +244,9 @@ class PlacementHandler:
             else:
                 edge_idx = self._find_nearest_edge(x, y)
                 if edge_idx is not None and self.state.edge_source is not None:
+                    # Capture state before deleting the edge.
+                    self._push_undo()
+
                     edge_data = {
                         key: list(value) for key, value in self.state.edge_source.data.items()
                     }
@@ -1990,6 +285,9 @@ class PlacementHandler:
             self.state.selected_node_index = None
             self._clear_temp_edges()
             self._clear_node_table()
+
+        # Capture state before placing a new component.
+        self._push_undo()
 
         icon_key = self.state.selected_component
         placement_count = self.state.placed_counter.get(icon_key, 0) + 1
